[BUILTIN] Stop using sysexits.h in commandcmd
This gets rid of the only reference of sysexits.h in dash which is
from commandcmd.  This is needed for klibc support since it doesn't
have sysexits.h.

The only uses of sysexits.h in commandcmd is superfluous anyway.  In
fact, it is overly sensitive about usages such as 'command -vV ls'.
By making its behaviour close to that of bash/ksh, we end up saving
a bit of space too.