[JOBS] Fix cmdtxt crash on if statements
(Herbert: for context, see
http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=467065 )

This is a real bug in upstream dash, which has existed since at
least dash-0.5.1 (July 2004).  It is a latent bug in cmdtxt()
(which I think applies only to pipes) as it handles "if" commands.
The attached patch fixes it for me.

It's possible this patch will fix one or more of #462414,
#462977, and #463649.  I'll send messages there to see if
the submitters can test.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>