[EXPAND] Fixed non-leading slash treatment in expmeta
Back in January an attempt was made to fix the interpretation of
quoted slashes in expmeta.  However, this only fixed those cases
where the quoted slash is at the front of the word.  The case of
non-leading slashes caused the previous directory part to gain a
back slash suffix which causes subsequent pattern matches to fail.

This patch fixes this by removing the back slash in that case.

Thanks to Romain Tartière fox reporting this bug.

Test case:

    echo /*"/null"

Old result:

    /*/null

New result:

    /dev/null

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>