[SHELL] Add preliminary LINENO support
Looks like in contrast to what the dash.1 manual page says, expansion
of PS{1,2,4} does work.

Here is a little patch to set LINENO. The ways in that it is less than
ideal mirror the ways that the line number error reporting is also
less than ideal.

For example if you run this:

(
    x=$((1/0))
    # Just to add another line
    # And another
) # error reports this line

The error reported will be  the closing parenthesis even though I
think most people would prefer the error to be the one where x was
set.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>