[ARITH] Fix binary operator parsing
Jilles Tjoelker reported that binary operator parsing doesn't
respect operator precedence correctly in the case where a lower-
precedence operator is followed by a higher-precedence operator,
and then by a lower-precedence operator.

This patch fixes this by stopping when we encounter a binary
oeprator with a precedence lower than one that we have already
encountered.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>