[JOBS] Fix for job control off warning
There seems to be a problem with the new version of dash
with job control off.  I can't tell if it is just a warning or is a
manifest bug.

usr/dash/trap.c: In function `exitshell':
usr/dash/trap.c:376: warning: suggest braces around empty body in an `if' statement

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>