[JOBS] Fix wait regression where it does not wait for all jobs
The sigsuspend patch broke wait by making it return after just
one job has completed.  This is because we rely on pendingsigs
to signal work and never clear it until waitcmd finishes.

This patch adds a separate gotsigchld for this purpose so we
can clear it before we start waiting.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>