[EVAL] Fixed trap/return regression due to SKIPEVAL removal
On Wed, Aug 11, 2010 at 08:06:16AM +0000, Guido Berhoerster wrote:
> 
> with the latest git version of dash trap actions are not
> evaluated in the context of a function.
> 
> The following script demonstrates the bug:
> ----8<----
> read_timeout () {
>     saved_traps="$(trap)"
>     trap 'printf "timed out\n"; eval "${saved_traps}"; return' TERM
>     ( sleep $1; kill -TERM $$ ) >/dev/null 2>&1 &
>     timer_pid=$!
>     read $2
>     kill $timer_pid 2>/dev/null
> }
> 
> read_timeout 5 value
> printf "read \"%s\"\n" "${value:=default}"
> 
> ---->8----
> The return statement in the trap inside the read_timeout function
> does not return from the function but rather exits the script.
> 
> With dash 0.5.5.1 it works as expected.

This bug was caused by the SKIPEVAL removal.  When the SKIPEVAL
hack was added to improve set -e support in traps, dotrap was
changed to return whether set -e was detected.  After the removal
of SKIPEVAL, set -e is now handled through exraise.

However, dotrap still returned a value which is now incorrectly
used to trigger an exraise.

This patch removes the vestigial link between dotrap and exraise.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>