[MAN] Document optional open parenthesis for case patterns
While inspecting some dash scripts on my system, I was surprised to
see that some of them use an open parenthesis at the beginning of case
patterns while that's not mentioned in the manpage. Dash currently is
fine with and without that parenthesis (parser.c:413). The attached
patch documents this feature.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>