[BUILTIN] Merge SKIPFUNC/SKIPFILE and only clear SKIPFUNC when leaving dotcmd
Currently upon leaving a dotcmd the evalskip state is reset so
if a continue/break statement is used within a dot script it would
have no effect outside of the dot script.

This is inconsistent with other shells.

This patch is based on one by Jilles Tjoelker and only clears
SKIPFUNC when leaving a dot script.  As a result continue/break
will remain in effect.

It also merges SKIPFUNC/SKIPFILE as they have no practical difference.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>