[TRAP] Make sure evalskip is zero before running traps
As it is if dotrap is called with evalskip set to a nonzero value,
it'll try to execute any set traps.  The result is that the first
command in the first set trap will be executed while the rest of
the trap will be silently ignored due to evalskip.  This is highly
counterintuitive, even though both bash and ksh exhibit a similar
behaviour.

This patch fixes it by skipping trap processing if evalskip is set
on entry.  It also adds a dotrap call to the top of evaltree to
ensure that

	while continue; do
		continue;
	done

has a chance of running traps.

Finally the pendingsigs check is moved into dotrap for compactness.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>